requirements.txt
fastapi
uvicorn
groq
pytesseract
pillow
python-multipart
PyPDF2